+++ otel/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## To be released
 
-- Add `otelmock` package that contains mocks required for unit testing
+- docs: Updated README.md and added examples with implemented code and unit tests
+- test: Add `oteltest` package that contains test helpers for unit testing
+- test: Add `otelmock` package that contains mocks required for unit testing
 
 ## v0.1.0